network requests

+++ app/src/main/java/com/ssessments/detail_news_fragment/DetailNewsViewModel.kt
@@ -3,7 +3,7 @@ package com.ssessments.detail_news_fragment
 import android.util.Log
 import androidx.lifecycle.ViewModel;
 
-class DetailNewsViewModel(newsID:Long) : ViewModel() {
+class DetailNewsViewModel(newsID:Int) : ViewModel() {
     // TODO: Implement the ViewModel
     init {
         Log.v("DETAILVIEWMODEL"," news id je $newsID")
